requirements.txt
flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
paramiko>=3.4.0
defusedxml>=0.7.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
